requirements.txt
python-dotenv
py-cord
Flask
python-dateutil